+++ peer/tworandomchoices/tworandomchoices.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 Uber Technologies, Inc.
+// Copyright (c) 2021 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -22,6 +22,7 @@ package tworandomchoices
 
 import (
 	"math/rand"
+	"sync"
 	"time"
 
 	"go.uber.org/yarpc/api/peer"
@@ -32,6 +33,8 @@ import (
 type twoRandomChoicesList struct {
 	subscribers []*subscriber
 	random      *rand.Rand
+
+	m sync.RWMutex
 }
 
 // Option configures the peer list implementation constructor.
@@ -58,6 +61,9 @@ func newTwoRandomChoicesList(cap int, source rand.Source) *twoRandomChoicesList 
 }
 
 func (l *twoRandomChoicesList) Add(peer peer.StatusPeer, _ peer.Identifier) abstractlist.Subscriber {
+	l.m.Lock()
+	defer l.m.Unlock()
+
 	index := len(l.subscribers)
 	l.subscribers = append(l.subscribers, &subscriber{
 		index: index,
@@ -67,6 +73,9 @@ func (l *twoRandomChoicesList) Add(peer peer.StatusPeer, _ peer.Identifier) abst
 }
 
 func (l *twoRandomChoicesList) Remove(peer peer.StatusPeer, _ peer.Identifier, ps abstractlist.Subscriber) {
+	l.m.Lock()
+	defer l.m.Unlock()
+
 	sub, ok := ps.(*subscriber)
 	if !ok || len(l.subscribers) == 0 {
 		return
@@ -79,6 +88,9 @@ func (l *twoRandomChoicesList) Remove(peer peer.StatusPeer, _ peer.Identifier, p
 }
 
 func (l *twoRandomChoicesList) Choose(_ *transport.Request) peer.StatusPeer {
+	l.m.RLock()
+	defer l.m.RUnlock()
+
 	numSubs := len(l.subscribers)
 	if numSubs == 0 {
 		return nil