+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - gRPC: accept dialer options in gRPC transportSpec config
+### Fixed
+- Http: fix memory and connection leak in http outbound call handler
 
 ## [1.49.1] - 2020-11-17
 ### Fixed