+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- gRPC: accept dialer options in gRPC transportSpec config
 ### Fixed
 - Http: fix memory and connection leak in http outbound call handler
 - Observability: error code CodeUnimplemented will be marked as an `application_error`
+- peer: the peer implementations provided by yarpc (hashring32, pendingheap,
+  randpeer, roundrobin, tworandomchoices) are thread-safe now
 
 ## [1.49.1] - 2020-11-17
 ### Fixed