+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.50.0] - 2021-01-22
 ### Added
 - gRPC: accept dialer options in gRPC transportSpec config
+
 ### Fixed
 - Http: fix memory and connection leak in http outbound call handler
 - Observability: error code CodeUnimplemented will be marked as an `application_error`